+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.10.4 (Unreleased)
 - **[Breaking]** Drop Ruby `3.0` support according to the EOL schedule.
+- [Fix] Toggle menu button post-turbo refresh stops working.
 
 ## 0.10.3 (2024-09-17)
 - **[Feature]** Introduce ability to brand Web UI with environment (Pro).