+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **[Breaking]** Drop Ruby `3.0` support according to the EOL schedule.
 - [Enhancement] Extract producers tracking `sync_threshold` into an internal config.
 - [Enhancement] Support complex Pro license loading strategies (Pro).
+- [Enhancement] Change default `retention.ms` for the metrics topic to support Redpanda Cloud defaults (#450).
 - [Enhancement] Include subscription group id in the consumers error tracking metadata.
 - [Enhancement] Collect metadata details of low level client errors when error tracking.
 - [Enhancement] Collect metadata details of low level listener errors when error tracking.